requirements.txt
CherryPy>=3.7.0
pyglib>=0.1
twitter.common.http>=0.3.3
twitter.common.exceptions>=0.3.3
twitter.common.zookeeper>=0.3.3
prometheus-client
